requirements.txt
feedparser==6.0.11
requests==2.31.0
flask==3.0.3
flask-cors==4.0.1
apscheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
pytz==2024.1
